+++ core/src/main/java/arez/annotations/Executor.java
@@ -13,7 +13,7 @@ public enum Executor
    */
   AREZ,
   /**
-   * The application is responsible for invoking the observed method via the {@link arez.ArezContext#track(Observer, arez.Function, Object...)}  or {@link arez.ArezContext#track(Observer, Procedure, Object...)} methods.
+   * The application is responsible for invoking the observed method via the {@link arez.ArezContext#observe(Observer, arez.Function, Object...)}  or {@link arez.ArezContext#observe(Observer, Procedure, Object...)} methods.
    */
   APPLICATION
 }