+++ doc-examples/src/main/java/arez/doc/examples/tracker/TrackerExample.java
@@ -16,14 +16,14 @@ public class TrackerExample
       //DOC ELIDE START
       //DOC ELIDE END
     };
-    final Observer tracker = Arez.context().tracker( () -> rescheduleRender() );
+    final Observer observer = Arez.context().tracker( () -> rescheduleRender() );
     //DOC ELIDE START
     //DOC ELIDE END
     // The rescheduleRender should ultimately result in the following
     // invocation. This call will need to be run at least once so that
     // the Arez runtime can determine the dependencies and reschedule
     // when the dependencies are changed.
-    Arez.context().track( tracker, observedFunction );
+    Arez.context().observe( observer, observedFunction );
   }
 
   static void rescheduleRender()